feat: Criando endpoints de remover tarefa

+++ ToDoList.Application/UseCases/TarefaUseCases/RemoverTarefaUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using ToDoList.Common;
+using ToDoList.Domain.Errors.Usuario;
 using ToDoList.Domain.Repositories;
 
 namespace ToDoList.Application.UseCases.TarefaUseCases;
@@ -12,8 +13,12 @@ public class RemoverTarefaUseCase
         _repository = repository;
     }
 
-    public async Task<Result> Execute(Guid tarefaId)
+    public async Task<Result> Execute(Guid usuarioId, Guid tarefaId)
     {
+        var tarefa = await _repository.BuscarTarefa(tarefaId);
+
+        if (tarefa.UsuarioId != usuarioId) return Result.Failure(UsuarioErrors.UsuarioProibidoDeRealizarAcao);
+
         await _repository.RemoverTarefa(tarefaId);
 
         return Result.Success();
